ingestion-service/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
python-dotenv
apscheduler